Splits core starter into autoconfigure and core module
