+++ spring-cloud-gcp-autoconfigure/pom.xml
@@ -10,6 +10,7 @@
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>spring-cloud-gcp-autoconfigure</artifactId>
+    <name>Spring Cloud GCP Autoconfigure Module</name>
 
     <dependencies>
         <dependency>
@@ -24,7 +25,7 @@
         </dependency>
         <dependency>
             <groupId>org.springframework.cloud</groupId>
-            <artifactId>spring-cloud-gcp-starter-core</artifactId>
+            <artifactId>spring-cloud-gcp-core</artifactId>
         </dependency>
         <dependency>
             <groupId>org.springframework.cloud</groupId>
