+++ spring-cloud-gcp-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,7 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+org.springframework.cloud.gcp.autoconfigure.core.GcpContextAutoConfiguration
 
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
 org.springframework.cloud.gcp.autoconfigure.config.GcpConfigAutoConfiguration,\
-org.springframework.cloud.gcp.core.autoconfig.GcpContextAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.core.GcpContextAutoConfiguration,\
 org.springframework.cloud.gcp.config.GoogleConfigPropertySourceLocator
