+++ spring-cloud-gcp-core/src/main/java/org/springframework/cloud/gcp/core/GcpProperties.java
