+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-trace/src/test/java/org/springframework/cloud/gcp/trace/autoconfig/StackdriverTraceAutoConfigurationTests.java
@@ -29,7 +29,7 @@ import org.junit.runner.RunWith;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.cloud.gcp.core.autoconfig.GcpContextAutoConfiguration;
+import org.springframework.cloud.gcp.autoconfigure.core.GcpContextAutoConfiguration;
 import org.springframework.cloud.sleuth.Span;
 import org.springframework.cloud.sleuth.autoconfig.TraceAutoConfiguration;
 import org.springframework.cloud.sleuth.log.SleuthLogAutoConfiguration;
@@ -48,7 +48,6 @@ import org.springframework.test.context.junit4.SpringRunner;
 		GcpContextAutoConfiguration.class,
 		TraceAutoConfiguration.class, SleuthLogAutoConfiguration.class }, properties = {
 				"spring.cloud.gcp.project-id=proj", "spring.sleuth.sampler.percentage=1.0" })
-
 public abstract class StackdriverTraceAutoConfigurationTests {
 
 	@Configuration
